requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
coverage==6.3
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
uuid==1.30
whitenoise==5.3.0
